requirements.txt
# Python 3.12+
fastapi>=0.109.0,<1.0.0
uvicorn[standard]>=0.27.0,<1.0.0
sqlalchemy>=2.0.25,<3.0.0
pydantic>=2.5.0,<3.0.0
pydantic-settings>=2.1.0,<3.0.0
python-telegram-bot>=21.0,<22.0
httpx>=0.26.0,<1.0.0

# Dev
pytest>=7.4.0,<9.0.0
